Move trait back for better diff for review
