+++ tests/ManagerTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Rbac\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Rbac\Tests\Common\ManagerTestTrait;
 
 final class ManagerTest extends TestCase
 {
